Move functions into separate file, add function keyword
The `peek()` function wasn't working properly. Clearly I
don't really need it, because I kind of forgot it's there,
but fixed it anyway.

As for `precmd()`, it never exhibited bugs, and `fps()`
is pretty useless though it worked, but I added the keyword.

Maybe I should figure this out, but meh.